Split interface/implementation for configuration set classes.
Refers #1362
