+++ restcomm/restcomm.commons/src/main/java/org/mobicents/servlet/restcomm/configuration/sets/CacheConfigurationSet.java
@@ -1,38 +1,33 @@
+/*
+ * TeleStax, Open Source Cloud Communications
+ * Copyright 2011-2014, Telestax Inc and individual contributors
+ * by the @authors tag.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation; either version 3 of
+ * the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>
+ *
+ */
+
 package org.mobicents.servlet.restcomm.configuration.sets;
 
-import org.mobicents.servlet.restcomm.configuration.sources.ConfigurationSource;
+/**
+ * @author orestis.tsakiridis@telestax.com - Orestis Tsakiridis
+ */
+public interface CacheConfigurationSet {
 
-public class CacheConfigurationSet extends ConfigurationSet {
-    public static final String CACHE_NO_WAV_KEY = "runtime-settings.cache-no-wav";
-    public static final String CACHE_PATH_KEY = "runtime-settings.cache-path";
-    public static final String CACHE_URI_KEY = "runtime-settings.cache-uri";
+    boolean isNoWavCache();
 
-    private boolean noWavCache;
-    private String cachePath;
-    private String cacheUri;
+    String getCachePath();
 
-    public CacheConfigurationSet(ConfigurationSource source) {
-        super(source);
-
-        String value = source.getProperty(CACHE_NO_WAV_KEY);
-
-        // default flag value is "false" if appropriate key "cache-no-wav" is absent in configuration *.xml file
-        noWavCache = (value == null) ? false : Boolean.valueOf(source.getProperty(CACHE_NO_WAV_KEY));
-
-        cachePath = source.getProperty(CACHE_PATH_KEY);
-
-        cacheUri = source.getProperty(CACHE_URI_KEY);
-    }
-
-    public boolean isNoWavCache() {
-        return noWavCache;
-    }
-
-    public String getCachePath() {
-        return cachePath;
-    }
-
-    public String getCacheUri() {
-        return cacheUri;
-    }
+    String getCacheUri();
 }
