+++ restcomm/restcomm.commons/src/main/java/org/mobicents/servlet/restcomm/configuration/sets/MainConfigurationSet.java
@@ -20,100 +20,23 @@
 
 package org.mobicents.servlet.restcomm.configuration.sets;
 
-import org.mobicents.servlet.restcomm.annotations.concurrency.Immutable;
-import org.mobicents.servlet.restcomm.configuration.sources.ConfigurationSource;
 import org.mobicents.servlet.restcomm.http.SslMode;
-import org.apache.commons.lang.StringUtils;
 
 /**
- * Provides a typed interface to a set of configuration options retrieved from a
- * configuration source.
- *
- * To add a new option in this set define its name as static fields and then initialize,
- * validate it in the constructor.
- *
- * @author orestis.tsakiridis@telestax.com (Orestis Tsakiridis)
- *
+ * @author orestis.tsakiridis@telestax.com - Orestis Tsakiridis
  */
-@Immutable
-public class MainConfigurationSet extends ConfigurationSet {
+public interface MainConfigurationSet {
+    SslMode getSslMode();
 
-    private static final String SSL_MODE_KEY = "http-client.ssl-mode";
-    private static final String HTTP_RESPONSE_TIMEOUT = "http-client.response-timeout";
-    private static final SslMode SSL_MODE_DEFAULT = SslMode.strict;
-    private final SslMode sslMode;
-    private final int responseTimeout;
-    private static final String USE_HOSTNAME_TO_RESOLVE_RELATIVE_URL_KEY = "http-client.use-hostname-to-resolve-relative-url";
-    private static final String HOSTNAME_TO_USE_FOR_RELATIVE_URLS_KEY = "http-client.hostname";
-    private static final boolean RESOLVE_RELATIVE_URL_WITH_HOSTNAME_DEFAULT = true;
-    private final boolean useHostnameToResolveRelativeUrls;
-    private final String hostname;
-    private String instanceId;
+    int getResponseTimeout();
 
-    public static final String BYPASS_LB_FOR_CLIENTS = "bypass-lb-for-clients";
-    private boolean bypassLbForClients = false;
+    boolean isUseHostnameToResolveRelativeUrls();
 
-    public MainConfigurationSet(ConfigurationSource source) {
-        super(source);
-        SslMode sslMode;
-        boolean resolveRelativeUrlWithHostname;
-        String resolveRelativeUrlHostname;
-        boolean bypassLb = false;
-        int timeout = 5000;
+    String getHostname();
 
-        try {
-            timeout = Integer.parseInt(source.getProperty(HTTP_RESPONSE_TIMEOUT));
-        } catch (Exception e) {
-            throw new RuntimeException("Error initializing '" + HTTP_RESPONSE_TIMEOUT + "' configuration setting", e);
-        }
-        // http-client.ssl-mode
-        try {
-            sslMode = SSL_MODE_DEFAULT;
-            String sslModeRaw = source.getProperty(SSL_MODE_KEY);
-            if ( ! StringUtils.isEmpty(sslModeRaw) )
-                sslMode = SslMode.valueOf(sslModeRaw);
-        } catch (Exception e) {
-            throw new RuntimeException("Error initializing '" + SSL_MODE_KEY  + "' configuration setting", e);
-        }
-        this.sslMode = sslMode;
+    boolean getBypassLbForClients();
 
-        // http-client.hostname
-        // http-client.use-hostname-to-resolve-relative-url
-        try {
-            resolveRelativeUrlWithHostname = RESOLVE_RELATIVE_URL_WITH_HOSTNAME_DEFAULT;
-            resolveRelativeUrlWithHostname = Boolean.valueOf(source.getProperty(USE_HOSTNAME_TO_RESOLVE_RELATIVE_URL_KEY));
-            resolveRelativeUrlHostname = source.getProperty("http-client.hostname");
-            bypassLb = Boolean.valueOf(source.getProperty(BYPASS_LB_FOR_CLIENTS));
-        } catch (Exception e) {
-            throw new RuntimeException("Error initializing '" + USE_HOSTNAME_TO_RESOLVE_RELATIVE_URL_KEY + "' configuration setting", e);
-        }
-        this.responseTimeout = timeout;
-        this.useHostnameToResolveRelativeUrls = resolveRelativeUrlWithHostname;
-        this.hostname = resolveRelativeUrlHostname;
-        bypassLbForClients = bypassLb;
-    }
+    void setInstanceId(String instanceId);
 
-    public SslMode getSslMode() {
-        return sslMode;
-    }
-
-    public int getResponseTimeout() {
-        return responseTimeout;
-    }
-
-    public boolean isUseHostnameToResolveRelativeUrls() {
-        return useHostnameToResolveRelativeUrls;
-    }
-
-    public String getHostname() {
-        return hostname;
-    }
-
-    public boolean getBypassLbForClients() { return bypassLbForClients; }
-
-    public void setInstanceId(String instanceId) {
-        this.instanceId = instanceId;
-    }
-
-    public String getInstanceId() { return this.instanceId; }
+    String getInstanceId();
 }
