+++ restcomm/restcomm.commons/src/test/java/org/mobicents/servlet/restcomm/configuration/sets/CustomConfigurationSet.java
@@ -1,5 +1,6 @@
 package org.mobicents.servlet.restcomm.configuration.sets;
 
+import org.mobicents.servlet.restcomm.configuration.sets.impl.ConfigurationSet;
 import org.mobicents.servlet.restcomm.configuration.sources.ConfigurationSource;
 
 public class CustomConfigurationSet extends ConfigurationSet {
